requirements.txt
torch==2.0.1
torchvision==0.15.2
numpy==1.25.1
pandas==2.0.3
opencv-python==4.8.0.74
Pillow==10.0.0
pyyaml==6.0
flake8==6.0.0
black==23.3.0
seaborn==0.13.0
scikit-image==0.21.0
scikit-learn==1.3.0
matplotlib==3.7.1
